--- src/gameskyrim.json
@@ -1,1 +1,0 @@
-{}
