+++ src/skyrimSEbsainvalidation.cpp
@@ -1,0 +1,16 @@
+#include "skyrimSEbsainvalidation.h"
+
+SkyrimSEBSAInvalidation::SkyrimSEBSAInvalidation(DataArchives *dataArchives, MOBase::IPluginGame const *game)
+  : GamebryoBSAInvalidation(dataArchives, "skyrim.ini", game)
+{
+}
+
+QString SkyrimSEBSAInvalidation::invalidationBSAName() const
+{
+  return "Skyrim - Invalidation.bsa";
+}
+
+unsigned long SkyrimSEBSAInvalidation::bsaVersion() const
+{
+  return 0x68;
+}
