+++ src/skyrimSEsavegameinfo.h
